test: share the XR-1 policy fixtures through conftest
The initialized-policy fixtures - the tiny kwargs, the dataset statistics and
the offline backbone that keeps construction off the Hub - lived in
test_policy.py, so a second test module could not build a policy without
copying them.

Move them to conftest.py, which is where shared fixtures belong, and let
test_policy.py take the dataset widths from the constants already defined
there instead of restating them.
